refactoring plugin manager
