+++ gpt_automation/impl/plugin_impl/plugin_config.py
